rename skip_on_windows -> SkipOnWindows to follow naming convention
