+++ integration-tests/__tests__/babel_plugin_jest_hoist.test.js
@@ -9,7 +9,7 @@
 'use strict';
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {run} = require('../utils');
 const runJest = require('../runJest');
