+++ integration-tests/__tests__/compare_dom_nodes.test.js
@@ -8,7 +8,7 @@
  */
 'use strict';
 
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const runJest = require('../runJest');
 
 skipOnWindows.suite();
