+++ integration-tests/__tests__/console_log_output_when_run_in_band.test.js
@@ -10,7 +10,7 @@
 'use strict';
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {extractSummary, cleanup, writeFiles} = require('../utils');
 const runJest = require('../runJest');
