+++ integration-tests/__tests__/coverage_report.test.js
@@ -10,7 +10,7 @@
 
 const fs = require('fs');
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {extractSummary} = require('../utils');
 const runJest = require('../runJest');
