+++ integration-tests/__tests__/custom_matcher_stack_trace.test.js
@@ -10,7 +10,7 @@
 
 const runJest = require('../runJest');
 const {extractSummary} = require('../utils');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 
 skipOnWindows.suite();
