+++ integration-tests/__tests__/custom_reporters.test.js
@@ -8,7 +8,7 @@
  */
 'use strict';
 
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {cleanup, extractSummary, writeFiles} = require('../utils');
 const runJest = require('../runJest');
 const os = require('os');
