+++ integration-tests/__tests__/failures.test.js
@@ -8,7 +8,7 @@
  */
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {extractSummary} = require('../utils');
 const runJest = require('../runJest');
