+++ integration-tests/__tests__/force_exit.test.js
@@ -14,7 +14,7 @@ import os from 'os';
 import path from 'path';
 const {cleanup, writeFiles} = require('../utils');
 
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const DIR = path.resolve(os.tmpdir(), 'force_exit_test');
 
 skipOnWindows.suite();
