+++ integration-tests/__tests__/jest.config.js.test.js
@@ -10,7 +10,7 @@
 'use strict';
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {extractSummary, cleanup, writeFiles} = require('../utils');
 const runJest = require('../runJest');
