+++ integration-tests/__tests__/jest_require_mock.test.js
@@ -11,7 +11,7 @@
 
 const path = require('path');
 const os = require('os');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {cleanup, writeFiles} = require('../utils');
 const runJest = require('../runJest');
