+++ integration-tests/__tests__/multi_project_runner.test.js
@@ -14,7 +14,7 @@ import os from 'os';
 import path from 'path';
 
 const {cleanup, extractSummary, writeFiles} = require('../utils');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const DIR = path.resolve(os.tmpdir(), 'multi_project_runner_test');
 
 skipOnWindows.suite();
