+++ integration-tests/__tests__/native_async_mock.test.js
@@ -10,7 +10,7 @@
 'use strict';
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {run, extractSummary} = require('../utils');
 const runJest = require('../runJest');
