+++ integration-tests/__tests__/regex_(char_in_path.test.js
@@ -8,7 +8,7 @@
  */
 'use strict';
 
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const runJest = require('../runJest');
 
 describe('Regex Char In Path', () => {
