+++ integration-tests/__tests__/test_path_pattern_reporter_message.test.js
@@ -13,7 +13,7 @@ import os from 'os';
 import path from 'path';
 const {cleanup, writeFiles} = require('../utils');
 
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const DIR = path.resolve(os.tmpdir(), 'jest_path_pattern_reporter_message');
 
 skipOnWindows.suite();
