+++ integration-tests/__tests__/transform.test.js
@@ -8,7 +8,7 @@
  */
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {
   run,
   cleanup,
