+++ integration-tests/__tests__/typescript_coverage.test.js
@@ -8,7 +8,7 @@
  */
 
 const path = require('path');
-const skipOnWindows = require('../../scripts/skip_on_windows');
+const skipOnWindows = require('../../scripts/SkipOnWindows');
 const {run} = require('../utils');
 const runJest = require('../runJest');
