+++ packages/jest-cli/src/__tests__/search_source.test.js
@@ -12,7 +12,7 @@ import path from 'path';
 
 jest.setTimeout(15000);
 
-const skipOnWindows = require('../../../../scripts/skip_on_windows');
+const skipOnWindows = require('../../../../scripts/SkipOnWindows');
 
 const rootDir = path.resolve(__dirname, 'test_root');
 const testRegex = path.sep + '__testtests__' + path.sep;
