+++ packages/jest-haste-map/src/__tests__/index.test.js
@@ -93,7 +93,7 @@ jest.mock('graceful-fs', () => ({
 }));
 jest.mock('fs', () => require('graceful-fs'));
 
-const skipOnWindows = require('../../../../scripts/skip_on_windows');
+const skipOnWindows = require('../../../../scripts/SkipOnWindows');
 
 const cacheFilePath = '/cache-file';
 let consoleWarn;
