+++ packages/jest-haste-map/src/crawlers/__tests__/node.test.js
@@ -8,7 +8,7 @@
 
 'use strict';
 
-const skipOnWindows = require('../../../../../scripts/skip_on_windows');
+const skipOnWindows = require('../../../../../scripts/SkipOnWindows');
 
 jest.mock('child_process', () => ({
   spawn: jest.fn((cmd, args) => {
