+++ packages/jest-runtime/src/__tests__/runtime_cli.test.js
@@ -10,7 +10,7 @@
 
 const path = require('path');
 const {sync: spawnSync} = require('cross-spawn');
-const skipOnWindows = require('../../../../scripts/skip_on_windows');
+const skipOnWindows = require('../../../../scripts/SkipOnWindows');
 
 skipOnWindows.suite();
